op: 拆分代码
